Don't refer to biblionumber as Item Number
There are three places where a biblionumber is given the heading
"Item Number", which is confusing. This patch replaces "Item
Number" with "Biblio Number". This is still in violation of Koha
coding guideline "HTML4", which says that only first letters
should be uppercase:
http://wiki.koha-community.org/wiki/Coding_Guidelines#HTML4:_Upper_and_lower_cases_in_strings
But changing all the strings is something of a big task, so this
patch chooses to reduce confusion while being consistent with
the rest of the ILL work.