WI EDI: Handle 'Howes' QUOTES with 'PRI+AAB'
EDI assumes that quotes messages contain a PRI+AAE block. This is not
the case for Howes where the PRI block is an AAB form.

This patch adds handling to EDI.pm to check for the expected AAE block
first and then fallback to AAB if it is undefined.

The symptom was very wierd: As the ->price_info method yielded an uninitialized
value then the hasref in which it was being called would end up with a random
key/value transposition during instantiation which would lead to a database column
error later in processing which was not caught in any meaningful way and prevented
further processing.