Bug 24160: Fix OPACHoldsIfAvailableAtPickup to follow holds policy
Bug 17453 added the ability to restrict the pick up location for items,
if at least one item was available at this location.
It assumed that an available item was an item that was not checked out,
not lost and not damaged.
But actually we should follow holds policy, defined in the circulation
rules.
An item that is not checked out but on which no hold can be placed on,
it should be considered as not available.

Test plan:
0/ Setting up the problematic situation
Considering a bibliographic record with 3 items I1, I2, and I3:
  I1, I2 and I3 have the same location (say Centerville)
  I1 and I2 have the item type 'Book' and are checked out
  I3 is a "Computer Files"
"Computer Files" has a "No holds allowed" hold policy (bottom of the
circulation rules view)

Turn on off (Don't allow) OPACHoldsIfAvailableAtPickup

1/ At the OPAC place a hold on this bibliographic record (biblio or
item-level hold), open the "Pick up location" dropdown list
=> Without this patch you cannot select "Centerville"
=> With this patch applied you can select it

2/ Confirm the hold
=> The hold as been correctly placed on the item and the pickup location
is correct.

3/ Cancel the hold
4/ Check I1 in

5/ Repeat 1/
=> Without and with this patch you cannot select "Centerville"

6/ Open the HTML inspector in the browser, select the dropdown list and
remove the 'disabled="disabled"' attribute of the CPL option

7/ Select (force) Centerville

8/ Confirm the hold
=> Nothing happened! You tried to cheat and place a hold to pickup at an
invalid location.

Note for QA: I have tried to move that code to a module but it's not
trivial at all. The availability needs to be indexed in the search
engine to make all of this much more easy.

Sponsored-by: Educampus

Signed-off-by: Simon Perry <simon.perry@itcarlow.ie>