Bug 24080: (RM followup) Boraccount payout form correction
This patch adds a minor correction to the payout form on the borrower
acocunts page. It removes the hidden 'registerid' form field and
corrects the select field to submit registerid as intended.

This corrects the behaviour such that the staff user can select which
cash register the payout action should be attributed to.

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>