Reimplementation of b4d3d26c16b94ad33966683f0f08880807237bb2
The previous commit didn't survive the move to 19.11, so it's reimplmented here. Also it seems that at some point we moved to using the authorised value code rather than id in status alias, so this patch also makes that change in our filter

Fix hiding ILL requests filter

We're not retrieving certain ILL requests as a performance measure. For
some reason that is still a mystery to me, this used to work, but then
didn't. The query to retrieve requests should never have worked as "NOT
IN" should exclude rows where the column in question is null, but
somehow it did work. Then it didn't. I have now modified the query to
fix it.

Modify ILL request filters & conditionally filter

I have modified the ILL request statuses that can be displayed in the
main ILL list page, as per ticket
https://helpdesk.ptfs-europe.com/Ticket/Display.html?id=30736

I have also made a small modification to ensure that when viewing a
single patron's ILL request history, all requests are seen, not just
those matching the status whitelist