Bug 24190: (follow-up) Cancel order del item bug
If an order is cancelled but the associated bib / items are unable to be
removed, we go into error handling. We still need to log the
cancellation though. So this fix moves the logging to just after the
cancellation, so it will be logged regardless of the outcome with
associated things.

(cherry picked from commit 994950b0de3995d2c611b8b26383787e487c5ba7)