Bug 26041: Remove readonly flag
This patch removes the readonly attribute from the date fields for
'reserve suspension' on the holds table and in opac-user.

Note: I could not find where the readonly attribute is added to the date
fields used when places a reserver.

Note: I'm not sure about this followup.. if you enter a date in the
wrong format (American vs British for example) you might end up with a
bad date being submitted which could either result in a nasty server
error or the wrong date being used.