Bug 24975: Use 'key' for translation lookup
This patch switches from using the original text as a key for
translations to using a dedicated 'key' field.  We also add in a 'fuzzy'
flag to denote when the source text has changed and as such the
pre-existing translations may be incorrect.

ToDo:
1/ Expose 'Fuzzy' on the 'Localization' admin page
2/ Add a 'translation_key' method to Koha::Itemtype to encourage the
adoption of a standardly names accessor to get the 'translation_key'
from a Koha::Object derived class.