Bug 27378: Add JS to power modaljs
This commit adds the Javascript that provides the functionality within
the consentJS modaljs modal. It provides the ability to display the
current state of any consentJS values and to amend them, as well as add
new ones.

There is also some refactoring of the existing modalselect functions to
allow re-use of existing code between modalselect & modaljs. This
refactoring has been minimal and consists of abstracting functionality
into functions to allow more wide use.

Styles for the consentJS UI were also added.

Signed-ff-by: Barry Cannon <bc@interleaf.ie>