Bug 27378: Delete cookies when consent removed
We have provided the ability for a user to remove previously granted
consent, but we were not removing the cookie associated with that
consent. This was a complete oversight on my part.

I've introduced an additional bit of metadata for each "consent" defined
in the ConsentJS syspref, this allows the admin defining a "consent" to
specify a string (which is treated as a regex) that allows us to
identify a cookie name associated with that consent. So, for example, if
Google Analytics was being added, an identifying string might be '_ga'.
This allows us to subsequently remove a cookie that a previously
granted consent has set.

This commit also adds some missing 'html' filters in
koha-tmpl/intranet-tmpl/prog/en/includes/intranet-bottom.inc and
koha-tmpl/opac-tmpl/bootstrap/en/includes/opac-bottom.inc

Signed-ff-by: Barry Cannon <bc@interleaf.ie>
(cherry picked from commit dcd1ff6a2913c10d1ceb1b129a4d17a9a1efda6f)