Bug 27378: Add cookie domain and path
A subtlety I missed was that some tracking services, such as Google
Analytics, specify a cookie domain and path that differ from the
default. E.g. on a site with a domain of demostaff.koha-ptfs.co.uk, the
GA cookie will have a domain of '.koha-ptfs.co.uk' and a path of '/',
even though the cookie may have been set at a different path.

Both of these things makes sense, but of course unless we also specify
both of those things when we try to delete a previously set cookie, we
cannot.

Therefore, I have added two more bits of metadata to each "consent",
Cookie Domain and Cookie Path. It is necessary for an admin creating a
"consent" entry to know and specify these values. They can be
established by looking a cookie set by the service being configured.

This adds more complexity to setting up a consent, but we cannot remove
previously set cookies without it.

Signed-ff-by: Barry Cannon <bc@interleaf.ie>