Bug 29422: Add `before`, `after` and `between` match options
This patch adds three new comparison operators to the datatables api
wrapper.  The allow for date (and numberic) type comparisons in
searches.

The special case `between` operator expects a string value of the form
`start:end` and will fall back to `after` comparisons if only a start
number/date is passed.