Bug 31123: Add all subfields to OPAC stylsheet
This patch updates the OPAC XSLT's to display all subfields, barring the
'x' field which is often used as a 'private note' in other notes fields.

We also handle converting subfield 'u' to a hyper link if it's found to
be present and use subfield 'a' for the link text if present.

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
(cherry picked from commit 1c85b3052863ecb0ce94567b0b04d0705d6a514e)
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>