Bug 33484: Make saveState column handling global
This patch updates the handling for saveState added in bug 22276 to work
globally for all KohaTables that pass a table_settings hash.

Bug 33484: Revert template change introduced in bug 22276

With the globalisation of the stateSave handling introduced by bug 22276
we can remove the localised changes to the template in branches.tt

Bug 33484: Revert template change on suggestion.tt

This reverts commit 2b0693dbfd6aa2dd2c7c91428b23aba717ccc0af as it's no
longer required once we globalise the logic as we do in this patchset

Bug 33484: Add support to kohaTable (AJAX Tables)

Bug 33484: Schema Update

Bug 33484: Add save state option to table settings

Amended-by: Jonathan Druart
Drop changes to admin/columns_settings.yml to avoid conflict. Just
assume that 1 is the default (see next patch)

Bug 33484: Make default_save_state default to 1

Bug 33484: Add table settings to pending orders

Bug 33484: Clearly separate search and configuration

We want to make distinguish the search terms and the
configuration/option of the table (number of entries, column visibility,
etc.)

Bug 33484: Add tables_settings.default_save_state_search

Bug 33484: DBIC Schema change

Bug 33484: Implement Vue KohaTable

Move up as beforeMount was called too late to init DT's options

Bug 33484: Always save but restore on demand only

And remove get_columns_saved_state. I am not sure this what really
helpful, maybe in case the yml or config in the DB changed, but I don't
think we should complicate the code. If something is wrong, logout to
clear localStorage...
We will see later if we really need it.

Bug 33484: (Bug 27467) Ability to share a link

Bug 33484: (Bug 27467) Fix patrons search

Bug 33484: Remove localStorage on logout

Bug 33484: Redraw columns filters when column visibility is modified

This seems to be the most sensible approach, we redraw the whole th when
the column visibility is modified.

Bug 33484: Fix patrons search

Setting a DT's custom setting to know if we loaded an existing state, so
that we do not pick the value of the selects from the form (on the
left).

Bug 33484: Fix patron search form

Reset DT's filters if we are searching again using the search form.

If you use the form on the left we need to reset the different filters
that could have been set from a previous state we loaded.

Bug 33484: Fix category select

Bug 33484: Do not defer loading if state is loaded from localStorage

Squashing
Revert "Bug 33484: (Bug 27467) Fix patrons search"
This reverts commit 0d21ff768b97bc8393a01b0ad881961a3252f2cf.

Bug 33484: Early return in add_filters if node does not exist

Maybe we could even return early from kohaTable?

Bug 33484: Adjust columns_settings.inc

Bug 33484: Fix pending orders

Sometimes 'this' is the DOM element, for instance on the pending orders
table.

Bug 33484: Fix Vue tables

datatables.net-vue3 does not use a specific id, and does not let us pass
one. It is generated, like: DataTables_Table_0 (and increment when a new
table is created).
So it's not possible to use this id: it can change and so is not
reliable.

I have tried to pass "data-koha-table-id" with a specific id
(KohaTable_agreements). Which was working, but actually we could simply
use the table settings key to identify the tables in the URL.

Bug 33484: Fix bookings

Bug 33484: Fix libraries table

We added columns to the table but never displayed them, and they were
not part of the columns settings.
I am not sure what was the purpose of that but it looks completely
wrong. If we want to restore them it should be done correctly (ie. add
them back to the column list and add the new ones to the columns
settings).