Bug 33260: Move patron search into main body of page
This patch moves the patron search in the Notices management page to the main body of the page so that the patron search result table displays more nicely.

This is a better place for it anyway, as using this search takes the user to a different page. In the sidebar, where it was previously, it implied it would have the same results display as the search filters.

Signed-off-by: Laura Escamilla <laura.escamilla@bywatersolutions.com>
Signed-off-by: kelly <kelly@bywatersolutions.com>
Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>